fix(engine): agent cargo + float food_stock
Two linked schema fixes for the engine's float economics:

1. Agents carry a float pouch. Engine CARRY_MAX=8.0 and
   FORAGE_TILE_DEPLETION=2.0 mutate a Float field, but the DB had no
   column — worker restart dropped the cargo and foraged food
   vanished. Add agents.cargo Float NOT NULL DEFAULT 0.0.

2. colonies.food_stock was Integer while harvest/deposit/eat_from_cache
   wrote floats. Postgres truncated/rejected fractional increments.
   Widen Int -> Float with USING CAST; existing stock values preserved.

Schema, ORM models, and mappers land in a single commit so rows_to_world
rehydration stays consistent — the schema<->mapper invariant documented
in the Foundation plan.

+++ backend/app/models/colony.py
@@ -10,5 +10,5 @@ class Colony(db.Model):
     camp_x = db.Column(db.Integer, nullable=False)
     camp_y = db.Column(db.Integer, nullable=False)
     food_stock = db.Column(
-        db.Integer, nullable=False, default=0, server_default='0',
+        db.Float, nullable=False, default=0.0, server_default='0',
     )
